app/requirements.txt
PySide6
Pillow
numpy
onnxruntime-directml